requirements.txt
flask==3.0.0
flask-basicauth==0.2.0
gunicorn==21.2.0
pillow==10.4.0
requests==2.31.0
httpx==0.27.0
openai==1.61.1